pyproject.toml
[project]
name = "roleradar"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apify-client>=2.5.0",
    "google-genai>=1.69.0",
    "google-generativeai>=0.8.6",
    "mcp[cli]>=1.26.0",
    "openai>=2.30.0",
    "pymupdf>=1.27.2.2",
    "python-dotenv>=1.2.2",
    "reportlab>=4.4.10",
    "streamlit>=1.55.0",
]
